fix data loader

+++ src/data_loader/utils/__init__.py
@@ -126,10 +126,8 @@ WAIT_TIME = 0.05
 
 
 def lazy_cycle(iterable):
-    while True:
-        iter_init = iter(iterable)
-        for element in iter_init:
-            yield element
+    for element in iterable:
+        yield element
 
 
 def default_collate_fn(data_object_list):
@@ -175,7 +173,6 @@ class BaseProcessPool():
     def reset_states(self):
         self.batch_idx = 0
         self.shuffle_idx()
-        self.put_idx_to_queue()
 
     def shuffle_idx(self):
         if self.shuffle:
@@ -223,37 +220,78 @@ class MultiProcessPool(BaseProcessPool):
         self.shuffle = shuffle
         self.data_num = len(data_getter)
         self.idx_list = list(range(self.data_num))
+        self.process_list = None
         self.batch_num = math.ceil(self.data_num / batch_size)
         self.batch_size = batch_size
         self.num_workers = num_workers
         self.verbose = verbose
+        self.idx_queue = mp.Queue()
+        self.output_queue = mp.Queue()
+        self.reset_states()
+        self.put_idx_to_queue()
+        self.process_list = []
+        for _ in range(self.num_workers):
+            process = mp.Process(target=consumer_fn, args=(self.data_getter,
+                                                           self.idx_queue,
+                                                           self.output_queue,
+                                                           self.batch_size))
+            # process.daemon = True
+            process.start()
+            self.process_list.append(process)
 
     def __iter__(self):
-        self.start_process()
-        try:
-            while self.batch_idx < self.batch_num:
-                start_idx = self.batch_size * self.batch_idx
-                end_idx = min(start_idx + self.batch_size, self.data_num)
-                current_batch_size = end_idx - start_idx
-                data_object_list = []
-                for _ in range(current_batch_size):
-                    while len(data_object_list) < current_batch_size:
-                        try:
-                            batch_idx, data_object = self.output_queue.get(
-                                timeout=0)
-                            data_object_list.append(data_object)
-                            if self.verbose:
-                                print(batch_idx)
-                        except Empty:
-                            sleep(WAIT_TIME)
-                            continue
+        # self.start_process()
+        self.reset_states()
+        self.put_idx_to_queue()
 
-                batch_data_tuple = self.collate_fn(data_object_list)
+        while self.batch_idx < self.batch_num:
+            start_idx = self.batch_size * self.batch_idx
+            end_idx = min(start_idx + self.batch_size, self.data_num)
+            current_batch_size = end_idx - start_idx
+            data_object_list = []
+            for _ in range(current_batch_size):
+                while len(data_object_list) < current_batch_size:
+                    try:
+                        batch_idx, data_object = self.output_queue.get(
+                            timeout=0)
+                        data_object_list.append(data_object)
+                        if self.verbose:
+                            print(batch_idx)
+                    except Empty:
+                        sleep(WAIT_TIME)
+                        continue
 
-                self.batch_idx += 1
-                yield batch_data_tuple
-        finally:
-            self.terminate_process()
+            batch_data_tuple = self.collate_fn(data_object_list)
+
+            self.batch_idx += 1
+            yield batch_data_tuple
+        # self.terminate_process()
+
+    def __next__(self):
+        if self.batch_idx == 0:
+            self.reset_states()
+        if self.batch_idx == self.batch_num:
+            self.reset_states()
+            self.put_idx_to_queue()
+        if self.batch_idx < self.batch_num:
+            start_idx = self.batch_size * self.batch_idx
+            end_idx = min(start_idx + self.batch_size, self.data_num)
+            current_batch_size = end_idx - start_idx
+            data_object_list = []
+
+            while len(data_object_list) < current_batch_size:
+                try:
+                    batch_idx, data_object = self.output_queue.get(timeout=0)
+                    data_object_list.append(data_object)
+                    if self.verbose:
+                        print(batch_idx)
+                except Empty:
+                    sleep(WAIT_TIME)
+                    continue
+
+            batch_data_tuple = self.collate_fn(data_object_list)
+            self.batch_idx += 1
+            return batch_data_tuple
 
     def start_process(self):
         self.idx_queue = mp.Queue()
@@ -265,18 +303,24 @@ class MultiProcessPool(BaseProcessPool):
                                                            self.idx_queue,
                                                            self.output_queue,
                                                            self.batch_size))
-            process.daemon = True
+            # process.daemon = True
             process.start()
             self.process_list.append(process)
         if self.verbose:
             print("process_start")
+
     def terminate_process(self):
-        for process in self.process_list:
-            process.terminate()
-            process.join()
-        self.idx_queue.close()
-        self.output_queue.close()
-        if self.verbose:
-            print("process_terminate")
+        if self.process_list is not None:
+            for idx, process in enumerate(self.process_list):
+                process.terminate()
+                process.join()
+                print(f"process {idx} terminate")
+            del self.process_list
+            self.process_list = []
+            self.idx_queue.close()
+            self.output_queue.close()
+            if self.verbose:
+                print("process_terminate")
+
     def __len__(self):
         return self.batch_num
